[connectors] Refactor Redis client structure to centralize logic and streamline usage across modules.
