+++ connectors/src/connectors/notion/temporal/activities.ts
@@ -75,7 +75,6 @@ import {
   NotionDatabase,
   NotionPage,
 } from "@connectors/lib/models/notion";
-import { redisClient } from "@connectors/lib/redis";
 import { syncStarted, syncSucceeded } from "@connectors/lib/sync_status";
 import { heartbeat } from "@connectors/lib/temporal";
 import mainLogger from "@connectors/logger/logger";
@@ -92,6 +91,7 @@ import {
   isDevelopment,
   slugify,
 } from "@connectors/types";
+import { redisClient } from "@connectors/types/shared/redis";
 import { sha256 } from "@connectors/types/shared/utils/hashing";
 
 const logger = mainLogger.child({ provider: "notion" });
