+++ connectors/src/types/shared/rate_limiter.ts
@@ -1,39 +1,13 @@
-import type { LoggerInterface } from "@dust-tt/client";
-import type { Result } from "@dust-tt/client";
+import type { LoggerInterface, Result } from "@dust-tt/client";
 import { Err, Ok } from "@dust-tt/client";
 import { v4 as uuidv4 } from "uuid";
 
 import { normalizeError } from "@connectors/types";
 
-import type { RedisUsageTagsType } from "../shared/redis_client";
-import { redisClient } from "../shared/redis_client";
+import { redisClient } from "./redis";
 import { getStatsDClient } from "./statsd";
 
 export class RateLimitError extends Error {}
-
-let rateLimiterRedisClient: Awaited<ReturnType<typeof redisClient>> | undefined;
-
-async function getRedisClient({
-  origin,
-  redisUri,
-}: {
-  origin: RedisUsageTagsType;
-  redisUri?: string;
-}) {
-  const REDIS_URI = redisUri || process.env.REDIS_URI;
-  if (!REDIS_URI) {
-    throw new Error("REDIS_URI is not defined");
-  }
-
-  if (!rateLimiterRedisClient) {
-    rateLimiterRedisClient = await redisClient({
-      origin,
-      redisUri: REDIS_URI,
-    });
-  }
-
-  return rateLimiterRedisClient;
-}
 
 export const RATE_LIMITER_PREFIX = "rate_limiter";
 
@@ -41,7 +15,6 @@ const makeRateLimiterKey = (key: string) => `${RATE_LIMITER_PREFIX}:${key}`;
 
 interface RateLimiterOptionsBase {
   key: string;
-  redisUri?: string;
 }
 
 export async function rateLimiter({
@@ -49,7 +22,6 @@ export async function rateLimiter({
   maxPerTimeframe,
   timeframeSeconds,
   logger,
-  redisUri,
 }: {
   logger: LoggerInterface;
   maxPerTimeframe: number;
@@ -61,16 +33,15 @@ export async function rateLimiter({
   const redisKey = makeRateLimiterKey(key);
   const tags: string[] = [];
 
-  let redis: undefined | Awaited<ReturnType<typeof redisClient>> = undefined;
-  try {
-    redis = await getRedisClient({ origin: "rate_limiter", redisUri });
+  const redis = await redisClient({ origin: "rate_limiter" });
 
-    const zcountRes = await redis.zCount(
+  try {
+    const zCountRes = await redis.zCount(
       redisKey,
       new Date().getTime() - timeframeSeconds * 1000,
       "+inf"
     );
-    const remaining = maxPerTimeframe - zcountRes;
+    const remaining = maxPerTimeframe - zCountRes;
     if (remaining > 0) {
       await redis.zAdd(redisKey, {
         score: new Date().getTime(),
@@ -108,14 +79,11 @@ export async function rateLimiter({
 
 export async function expireRateLimiterKey({
   key,
-  redisUri,
 }: RateLimiterOptionsBase): Promise<Result<boolean, Error>> {
-  let redis: undefined | Awaited<ReturnType<typeof redisClient>> = undefined;
+  const redis = await redisClient({ origin: "rate_limiter" });
+  const redisKey = makeRateLimiterKey(key);
 
   try {
-    redis = await getRedisClient({ origin: "rate_limiter", redisUri });
-    const redisKey = makeRateLimiterKey(key);
-
     const isExpired = await redis.expire(redisKey, 0);
 
     return new Ok(isExpired);
